fixed security issue

+++ IAPHelper/NSString+Base64.h
@@ -10,6 +10,6 @@
 
 @interface NSString (Base64)
 
-+ (NSString *) base64StringFromData:(NSData *)data length:(int)length;
++ (NSString *) base64StringFromData:(NSData *)data length:(long)length;
 
 @end
