+++ IAPHelper/NSString+Base64.m
@@ -18,7 +18,7 @@ static char base64EncodingTable[64] = {
 @implementation NSString (Base64)
 
 
-+ (NSString *) base64StringFromData: (NSData *)data length: (int)length {
++ (NSString *) base64StringFromData:(NSData *)data length:(long)length {
     unsigned long ixtext, lentext;
     long ctremaining;
     unsigned char input[3], output[4];
